Refactor: Separate delta and beta reduction.
