+++ base/src/main/java/org/aya/core/visitor/Subst.java
@@ -107,7 +107,7 @@ public record Subst(
   }
 
   public @NotNull Restr<Term> restr(@NotNull TyckState state, @NotNull Restr<Term> restr) {
-    return Expander.restr(restr.fmap(t -> t.subst(this).normalize(state, NormalizeMode.WHNF)));
+    return DeltaExpander.restr(restr.fmap(t -> t.subst(this).normalize(state, NormalizeMode.WHNF)));
   }
 
   @Override public @NotNull Doc toDoc(@NotNull DistillerOptions options) {
