+++ base/src/main/java/org/aya/tyck/LittleTyper.java
@@ -5,7 +5,7 @@ package org.aya.tyck;
 import org.aya.core.def.Def;
 import org.aya.core.def.PrimDef;
 import org.aya.core.term.*;
-import org.aya.core.visitor.Expander;
+import org.aya.core.visitor.DeltaExpander;
 import org.aya.core.visitor.Subst;
 import org.aya.generic.Arg;
 import org.aya.generic.Constants;
@@ -36,8 +36,8 @@ public record LittleTyper(@NotNull TyckState state, @NotNull LocalCtx localCtx) 
         var callRaw = term(access.of()).normalize(state, NormalizeMode.WHNF);
         if (!(callRaw instanceof CallTerm.Struct call)) yield ErrorTerm.typeOf(access);
         var core = access.ref().core;
-        var subst = Expander.buildSubst(core.telescope(), access.fieldArgs())
-          .add(Expander.buildSubst(call.ref().core.telescope(), access.structArgs()));
+        var subst = DeltaExpander.buildSubst(core.telescope(), access.fieldArgs())
+          .add(DeltaExpander.buildSubst(call.ref().core.telescope(), access.structArgs()));
         yield core.result().subst(subst);
       }
       case FormTerm.Sigma sigma -> {
@@ -97,7 +97,7 @@ public record LittleTyper(@NotNull TyckState state, @NotNull LocalCtx localCtx) 
 
   private @NotNull Term defCall(@NotNull CallTerm.DefCall call) {
     return Def.defResult(call.ref())
-      .subst(Expander.buildSubst(Def.defTele(call.ref()), call.args()))
+      .subst(DeltaExpander.buildSubst(Def.defTele(call.ref()), call.args()))
       .lift(call.ulift());
   }
 }
