+++ base/src/main/java/org/aya/tyck/pat/PatTycker.java
@@ -22,8 +22,8 @@ import org.aya.core.def.Def;
 import org.aya.core.pat.Pat;
 import org.aya.core.pat.PatMatcher;
 import org.aya.core.term.*;
+import org.aya.core.visitor.DeltaExpander;
 import org.aya.core.visitor.EndoFunctor;
-import org.aya.core.visitor.Expander;
 import org.aya.core.visitor.Subst;
 import org.aya.generic.Constants;
 import org.aya.generic.util.InternalException;
@@ -401,7 +401,7 @@ public final class PatTycker {
   mischa(CallTerm.Data dataCall, CtorDef ctor, @Nullable LocalCtx ctx, @NotNull TyckState state) {
     if (ctor.pats.isNotEmpty()) return PatMatcher.tryBuildSubstTerms(ctx, ctor.pats, dataCall.args().view()
       .map(arg -> arg.term().normalize(state, NormalizeMode.WHNF)));
-    else return Result.ok(Expander.buildSubst(Def.defTele(dataCall.ref()), dataCall.args()));
+    else return Result.ok(DeltaExpander.buildSubst(Def.defTele(dataCall.ref()), dataCall.args()));
   }
 
   private record BodySubstitutor(
