+++ build.gradle.kts
@@ -175,6 +175,7 @@ subprojects {
           dev("dark-flames", "Darkflames", "dark_flames@outlook.com")
           dev("tsao-chi", "tsao-chi", "tsao-chi@the-lingo.org")
           dev("lunalunaa", "Luna Xin", "luna.xin@outlook.com")
+          dev("wsx", "Shuxian Wang", "wsx@berkeley.edu")
         }
         scm {
           connection.set("scm:git:$githubUrl")
